+++ CHANGELOG.md
@@ -1,8 +1,15 @@
 # MapTiler 3D Models Changelog
 
+## 2.0.1
+## Bug Fixes
+- Set default rendering min zoom value to work fine with globe 
+
 ## 2.0.0
-### Other
-- Update with MapTiler SDK v3 (globe)
+## New Features
+- Globe projection support
+
+## Others
+- Using MapTiler SDK JS v3 (globe)
 
 ## 1.0.0
 First release 🎉